Fix variables type in tar_rm_file function
Change file_start, file_end and tar_end to off_t type.

+++ src/main/tar_rm.c
@@ -81,9 +81,9 @@ static int tar_rm_file(int tar_fd, const char *filename)
 
   file_size = get_file_size(&file_header);
   
-  ssize_t file_start = lseek(tar_fd, -BLOCKSIZE, SEEK_CUR), // on était à la fin d'un header, on se place donc au début
-          file_end   = file_start + BLOCKSIZE + number_of_block(file_size)*BLOCKSIZE, 
-          tar_end    = lseek(tar_fd, 0, SEEK_END);
+  off_t file_start = lseek(tar_fd, -BLOCKSIZE, SEEK_CUR), // on était à la fin d'un header, on se place donc au début
+        file_end   = file_start + BLOCKSIZE + number_of_block(file_size)*BLOCKSIZE, 
+        tar_end    = lseek(tar_fd, 0, SEEK_END);
 
   if(fmemmove(tar_fd, file_end, tar_end - file_end, file_start) < 0)
     return -1;
